Fixes bug that fails to reset board/completes ui
Fixes a bug that fails to reset board upon sign out. This was done by
making a new game button that both shows the board upon a click
and resets all values of the array and respective divs on the
board. This completes all ui and game logic required for project.